% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionBttnPlus.R
\name{actionBttnPlus}
\alias{actionBttnPlus}
\title{Extend \code{shinyWidgets::actionBttn}}
\usage{
actionBttnPlus(inputId, label = NULL, icon = NULL, style = "unite",
  color = "default", size = "md", block = FALSE, no_outline = TRUE,
  classPlus = "", stylePlus = "", ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button, usually a text label.}

\item{icon}{An optional icon to appear on the button.}

\item{style}{Style of the button, to choose between \code{simple}, \code{bordered},
\code{minimal}, \code{stretch}, \code{jelly}, \code{gradient}, \code{fill},
\code{material-circle}, \code{material-flat}, \code{pill}, \code{float}, \code{unite}.}

\item{color}{Color of the button : \code{default}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{royal}.}

\item{size}{Size of the button : \code{xs},\code{sm}, \code{md}, \code{lg}.}

\item{block}{Logical, full width button.}

\item{no_outline}{Logical, don't show outline when navigating with
keyboard/interact using mouse or touch.}

\item{classPlus}{String; pass additional classes to the button tag}

\item{stylePlus}{String; pass additional styles to the button tag}

\item{...}{Pass additional arguments to the button tag}
}
\description{
Like \code{actionBttn} and even awesomer
}
\seealso{
\link[shinyWidgets]{downloadBttn}
}
\author{
Andrea Berardi \email{Andrea.Berardi@PAREXEL.com}
}
