% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{errorMessage}
\alias{errorMessage}
\title{Produces a sweetAlert-powered customized error message}
\usage{
errorMessage(session, text = "Incorrect input entered",
  title = "Input error")
}
\arguments{
\item{session}{The current shiny session object}

\item{text}{Body text for the pop-up message}

\item{title}{Title text for the pop-up message}
}
\value{
None
}
\description{
Produces a sweetAlert-powered customized error message
}
\examples{
errorMessage(session)

}
\author{
Andrea Berardi \email{Andrea.Berardi@PAREXEL.com}
}
