% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarPageLogo.R
\name{navbarPageLogo}
\alias{navbarPageLogo}
\title{Extends \code{shiny::navbarPage} to include logos easily}
\usage{
navbarPageLogo(title = NULL, ..., id = NULL, selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL, footer = NULL, inverse = FALSE,
  collapsible = FALSE, collapsable, fluid = TRUE, responsive = NULL,
  theme = NULL, windowTitle = title, logosHTML = NULL)
}
\arguments{
\item{title}{The title to display in the navbar}

\item{...}{\code{\link{tabPanel}} elements to include in the page. The
\code{navbarMenu} function also accepts strings, which will be used as menu
section headers. If the string is a set of dashes like \code{"----"} a
horizontal separator will be displayed in the menu.}

\item{id}{If provided, you can use \code{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link{tabPanel}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{position}{Determines whether the navbar should be displayed at the top
of the page with normal scrolling behavior (\code{"static-top"}), pinned at
the top (\code{"fixed-top"}), or pinned at the bottom
(\code{"fixed-bottom"}). Note that using \code{"fixed-top"} or
\code{"fixed-bottom"} will cause the navbar to overlay your body content,
unless you add padding, e.g.: \code{tags$style(type="text/css", "body
{padding-top: 70px;}")}}

\item{header}{Tag or list of tags to display as a common header above all
tabPanels.}

\item{footer}{Tag or list of tags to display as a common footer below all
tabPanels}

\item{inverse}{\code{TRUE} to use a dark background and light text for the
navigation bar}

\item{collapsible}{\code{TRUE} to automatically collapse the navigation
elements into a menu when the width of the browser is less than 940 pixels
(useful for viewing on smaller touchscreen device)}

\item{collapsable}{Deprecated; use \code{collapsible} instead.}

\item{fluid}{\code{TRUE} to use a fluid layout. \code{FALSE} to use a fixed
layout.}

\item{responsive}{This option is deprecated; it is no longer optional with
Bootstrap 3.}

\item{theme}{Alternative Bootstrap stylesheet (normally a css file within the
www directory). For example, to use the theme located at
\code{www/bootstrap.css} you would use \code{theme = "bootstrap.css"}.}

\item{windowTitle}{The title that should be displayed by the browser window.
Useful if \code{title} is not a string.}

\item{logosHTML}{A div including spans containing the logos to be displayed.
The div should be styled to float right and have the navbar-header and navbar-brand classes.
A valid logosHTML object is shown in the examples.}
}
\value{
Invisible TRUE if executes correctly, can be used for further validation or resets.
}
\description{
Extends \code{shiny::navbarPage} to include logos easily
}
\examples{
\dontrun{
logosHTML = div(
  class = "navbar-header navbar-brand",
  style = "float:right",
  span("", style = "display: block; overflow: auto; position: relative;"),
  span(img(
    src = "pathtofile.png",
    style = "height: 1em; vertical-align:center; margin-left: 10px;"
  ))
)
}
}
\seealso{
\code{\link[shiny]{navbarPage}}
}
\author{
Andrea Berardi \email{Andrea.Berardi@PAREXEL.com}
}
