% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlusPXL.R
\docType{package}
\name{shinyPlusPXL}
\alias{shinyPlusPXL}
\alias{shinyPlusPXL-package}
\title{shinyPlusPXL: A Package to Extend Shiny Functions for R/Shiny PAREXEL Applications}
\description{
shinyPlusPXL: A Package to Extend Shiny Functions for R/Shiny PAREXEL Applications
}
\author{
Andrea Berardi \email{Andrea.Berardi@PAREXEL.com}
}
