% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateThen.R
\name{validateThen}
\alias{validateThen}
\title{Extends \code{shiny::validate} to produce an informative pop-up message}
\usage{
validateThen(session, ..., errorTitle = NULL, raiseModal = TRUE)
}
\arguments{
\item{session}{The shiny client session.}

\item{...}{function calls to \code{\link[shiny]{validate}}.}

\item{errorTitle}{Optional error title string to be passed to \code{link{errorMessage}}.}

\item{raiseModal}{Logical, raise a modal \code{sweetAlert} if the input is invalid.}
}
\value{
Invisible TRUE if executes correctly, can be used for further validation or resets.
}
\description{
Extends \code{shiny::validate} to produce an informative pop-up message
}
\examples{
\dontrun{validateThen(session,
need(!is.na(x), "x is NA"),
errorTitle = "Invalid input", raiseModal = TRUE)}

}
\seealso{
\code{\link{validateThenCSS}}, \code{\link{errorMessage}},
\code{\link[shiny]{validate}}
}
\author{
Andrea Berardi \email{Andrea.Berardi@PAREXEL.com}
}
