% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateThenCSS.R
\name{validateThenCSS}
\alias{validateThenCSS}
\title{Extends \code{shiny::validate} to produce an informative pop-up message and applies a CSS class if invalid}
\usage{
validateThenCSS(session, inputId, ..., errorTitle = NULL,
  invalidCSS = "invalidInput", raiseModal = TRUE)
}
\arguments{
\item{session}{The shiny client session.}

\item{inputId}{The input id to be validated.}

\item{...}{function calls to \code{\link[shiny]{validate}}.}

\item{errorTitle}{Optional error title string to be passed to \code{link{errorMessage}}.}

\item{invalidCSS}{CSS class to be applied if invalid, or removed if valid.}

\item{raiseModal}{Logical, raise a modal \code{sweetAlert} if the input is invalid.}
}
\value{
Invisible TRUE if executes correctly, can be used for further validation or resets.
}
\description{
Extends \code{shiny::validate} to produce an informative pop-up message and applies a CSS class if invalid
}
\examples{
\dontrun{validateThenCSS(session,
"inputId", need(!is.na(x), "x is NA"))}

}
\seealso{
\code{\link{validateThen}}, \code{\link{errorMessage}},
\code{\link[shiny]{validate}}
}
\author{
Andrea Berardi \email{Andrea.Berardi@PAREXEL.com}
}
